feat: enable Base withdrawals (verified) + fully consolidate derivations into the registry
Base: missing from the curated withdraw gate since June — an
oversight, not a decision (Hugo). Verified live before enabling:
ARB→BASE quotes OK for ETH/USDC/USDT + outflow SDA create OK
(2026-07-13). Rollout-flagged (chain-rollout-base, created ON).

Full clean per review: chainRollout.consts and nonEvmWithdraw.consts
existed only to hold single derived constants — their exports now
live in chainRegistry.consts itself and the files are gone. No
re-export shims remain; rhino.consts/TokenSelector.consts keep their
derived chain maps because they co-locate with family-level constants
and a dozen consumers, but every value traces to one registry entry.
